refactor(devtools): modularize SVG converter and improve code generation
