+++ packages/nube-devtools/package.json
@@ -63,7 +63,6 @@
     "gulp": "^5.0.0",
     "gulp-zip": "^6.0.0",
     "postcss": "^8.5.3",
-    "prettier": "^3.0.3",
     "tailwindcss": "^4.1.3",
     "typescript": "^5.2.2",
     "vite": "^5.4.10"
